added delay analysis and get_latest_non_zero_delay method
